@@ -3,7 +3,7 @@ import 'package:bloc/bloc.dart';
 class SimpleBlocObserver extends BlocObserver {
   
   @override
-  void onEvent(Bloc bloc, Object event) {
+  void onEvent(Bloc bloc, Object? event) {
     super.onEvent(bloc, event);
     print(event);
   }
